# back-end/rest_api/src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:9115
    description: Generated server url
security:
  - observatoryAuth: []
paths:
...
  /api/logout:
    post:
      tags:
        - authentication-controller
      summary: Logout user
      description: Logs out the user by invalidating their token.
      operationId: logout
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
                example:
                  error: Invalid or expired token
        "200":
          description: Logout successful
  /api/login:
    post:
      tags:
        - authentication-controller
      summary: User Login
      description: Authenticate a user and return a token for accessing protected
        endpoints.
      operationId: login
      requestBody:
        description: User credentials
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Unauthorized401Response'
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: string
                example:
                  token: FOO
        "400":
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: string
                example:
                  error: Invalid credentials
  /api/admin/resetstations:
    post:
      tags:
        - admin-controller
      summary: Reset toll stations
      description: Resets the toll stations table with the data from the `tollstations2024.csv`
        file.
      operationId: resetStations
      responses:
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized401Response'
        "200":
          description: Reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetStations200Response'
        "400":
          description: Reset failed due to illegal argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetStations400Response'
        "500":
          description: Reset failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Generic500Response'
  /api/admin/resetpasses:
    post:
      tags:
        - admin-controller
      summary: Reset passes
      description: Resets the toll stations table with the data from the `tollstations2024.csv`
        file.
      operationId: resetPasses
      responses:
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized401Response'
        "200":
          description: Reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetStations200Response'
        "400":
          description: Reset failed due to illegal argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetStations400Response'
        "500":
          description: Reset failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Generic500Response'
  /api/admin/addpasses:
    post:
      tags:
        - admin-controller
      summary: Add passes
      description: Adds passes and updates data structures with the data from the
        provided csv file.
      operationId: addPasses
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized401Response'
        "200":
          description: Update successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetStations200Response'
        "500":
          description: Update failed due to interval server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Generic500Response'
        "400":
          description: Update failed due to illegal argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetStations400Response'
...
  /api/auth/operatorId:
    get:
      tags:
        - authentication-controller
      summary: Get Operator ID and Role
      description: Retrieves the operator ID (opId) and role of the logged-in user.
      operationId: getOperatorId
      responses:
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
                example:
                  error: Invalid or expired token
        "200":
          description: Operator ID and role retrieved successfully
          content:
            application/json:
              schema:
                type: string
                example:
                  opId: EG
                  role: admin
...
components:
  schemas:
    Unauthorized401Response:
      type: object
      properties:
        status:
          type: string
          example: failed
        info:
          type: string
          example: Invalid or expired token
      description: Unauthorized_401_response
    Generic500Response:
      type: object
      properties:
        status:
          type: string
          example: failed
        info:
          type: string
          example: Unexpected internal server error
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          description: User's username
          example: username1
        password:
          type: string
          description: User's password
          example: password1
    ResetStations200Response:
      type: object
      properties:
        status:
          type: string
          example: OK
    ResetStations400Response:
      type: object
      properties:
        status:
          type: string
          example: failed
        info:
          type: string
          example: "Invalid input: Check file structure or parameters"
...
  securitySchemes:
    observatoryAuth:
      type: apiKey
      name: X-OBSERVATORY-AUTH
      in: header